remove sqlite logic into a db_provider class
