+++ source/model/service/DatabaseManager.py
@@ -1,21 +1,11 @@
-import sqlite3
-from pathlib import Path
 from queue import Queue
 from PyQt6.QtCore import QObject, pyqtSignal, pyqtSlot
-import hashlib
-import secrets
-import json
 
-from model.data.document import Document, update_metadata_file
-from model.data.metadata import Metadata
+from model.service.db_providers import AbstractDatabaseProvider
+from model.data.document import Document
 
-VALID_FILTERS = {
-    'doc_id', 'metadata', 'deskewed', 'needs_approval', 'approved', 
-    'rejected', 'uploaded', 'path', 'metadata_file', 'metadata_file_type', 
-    'metadata_json', 'last_modified', 'error_msg'
-}
 
-class DatabaseManager(QObject):
+class DatabaseWorker(QObject):
     """
     A QObject worker that manages all database interactions
     in a separate thread.
@@ -24,23 +14,18 @@ class DatabaseManager(QObject):
     error = pyqtSignal(Exception,str) #error_msg, job_id
     prog = pyqtSignal(int)# progress
 
-    def __init__(self, db_path: Path, queue: Queue):
+    def __init__(self, provider: AbstractDatabaseProvider, queue: Queue):
         super().__init__()
-        self.db_path = db_path
+        self.provider = provider
         self.queue = queue
-        self.conn = None 
-        print('init DatabaseManager')
+        print('init DatabaseWanager')
 
     @pyqtSlot()
     def run(self):
         """The main worker loop. This runs in the QThread."""
         try:
             print('Database Manager queue running')
-            self.db_path.parent.mkdir(parents=True, exist_ok=True)
-            self.conn = sqlite3.connect(self.db_path)
-            self.conn.row_factory = sqlite3.Row
-            self.conn.execute("PRAGMA foreign_keys = ON;")
-            self._create_tables()
+            self.provider.connect()
 
             while True:
                 command, signals, data = self.queue.get()
@@ -53,24 +38,21 @@ class DatabaseManager(QObject):
                 try:
                     match command: 
                         case 'load_documents':
-                            filter_data = data
-                            docs = self._load_documents(filter_data)
+                            docs = self.provider.load_documents(data)
                             if not signals.is_cancelled():
                                 signals.data.emit(docs, signals.job_id)
                         case 'save_document':
-                            document = data
-                            self._save_document(document)
-                            signals.data.emit(document,signals.job_id)
-                            self.update.emit(document,signals.job_id)
+                            self.provider.save_document(data)
+                            signals.data.emit(data,signals.job_id)
+                            self.update.emit(data,signals.job_id)
                         case 'delete_document':
-                            doc_id = data
-                            self._delete_document(doc_id)
+                            self.provider.delete_document(data)
                             if not signals.is_cancelled():
-                                signals.data.emit(doc_id, signals.job_id)
-                                self.update.emit(doc_id, signals.job_id)
+                                signals.data.emit(data, signals.job_id)
+                                self.update.emit(data, signals.job_id)
                         case 'verify_login':
                             username, password = data
-                            success, role = self._verify_login(username, password)
+                            success, role = self.provider.verify_login(username, password)
                             
                             if not signals.is_cancelled():
                                 result = {'success': success, 'role': role, 'username': username}
@@ -78,327 +60,42 @@ class DatabaseManager(QObject):
                         case 'new_user':
                             username, password, role = data
                             if not signals.is_cancelled():
-                                success, msg = self._create_user(username, password, role)
+                                success, msg = self.provider.create_user(username, password, role)
                                 if success:
                                     signals.data.emit(True, signals.job_id)
                                 else:
                                     signals.error.emit(msg, signals.job_id)
                         case 'get_users':
                             if not signals.is_cancelled():
-                                users = self._get_users()
+                                users = self.provider.get_users()
                                 signals.data.emit(users, signals.job_id)
                         case 'delete_user':
-                            user_id = data
                             if not signals.is_cancelled():
-                                success, msg = self._delete_user(user_id)
+                                success, msg = self.provider.delete_user(data)
                                 if success:
                                     signals.data.emit(True, signals.job_id)
-
+                                else:
+                                    signals.error.emit(msg, signals.job_id)
                         case 'check_admin':
                             if not signals.is_cancelled():
-                                signals.data.emit(self._has_admin(), signals.job_id)
+                                signals.data.emit(self.provider.has_admin(), signals.job_id)
 
 
-                except sqlite3.OperationalError as e:
+                except Exception as e:
                     if "interrupted" in str(e).lower():
                         print(f"Database query for {signals.job_id} was successfully aborted.")
                         signals.error.emit("Cancelled by user", signals.job_id)
                     else:
-                        signals.error.emit(f"SQLite Error: {e}", signals.job_id)
-                
-                except Exception as e:
-                    signals.error.emit(e,signals.job_id)
-                    self.error.emit(e,signals.job_id)
+                        signals.error.emit(str(e), signals.job_id)
+                        self.error.emit(e, signals.job_id)
 
         except Exception as e:
             self.error.emit(e,'')
         finally:
-            if self.conn:
-                self.conn.close()
-
-    def ensure_table_schema(conn: sqlite3.Connection, table_name: str, primary_key:str,table_schema):
-        cursor = self.conn.cursor()
-        
-        cursor.execute(f"""
-            CREATE TABLE IF NOT EXISTS {table_name} (
-                {primary_key} TEXT PRIMARY KEY
-            )
-        """)
-        
-        # 3. Query the current state of the table
-        cursor.execute(f"PRAGMA table_info({table_name})")
-        existing_columns = {row[1] for row in cursor.fetchall()}
-        
-        # 4. Find and add any missing columns
-        for col_name, col_type in table_schema.items():
-            if col_name not in existing_columns:
-                print(f"Upgrading database schema: Adding column '{col_name}'...")
-                try:
-                    cursor.execute(f"ALTER TABLE {table_name} ADD COLUMN {col_name} {col_type}")
-                except sqlite3.OperationalError as e:
-                    print(f"Failed to add column {col_name}: {e}")
-
-        conn.commit()
-
-    def _create_tables(self):
-        """Internal method to create tables. Called by run()."""
-        cursor = self.conn.cursor()
-        cursor.execute('''
-            CREATE TABLE IF NOT EXISTS documents (
-                doc_id TEXT PRIMARY KEY,
-                metadata INTEGER NOT NULL DEFAULT 0,
-                deskewed INTEGER NOT NULL DEFAULT 0,
-                needs_approval INTEGER NOT NULL DEFAULT 0,
-                approved INTEGER NOT NULL DEFAULT 0,
-                rejected INTEGER NOT NULL DEFAULT 0,
-                uploaded INTEGER NOT NULL DEFAULT 0,
-                path TEXT,
-                metadata_file TEXT,
-                metadata_file_type TEXT,
-                metadata_json TEXT,
-                last_modified TEXT NOT NULL,
-                error_msg TEXT
-            )
-        ''')
-
-        cursor.execute('''
-            CREATE TABLE IF NOT EXISTS users (
-            user_id INTEGER PRIMARY KEY AUTOINCREMENT,
-            username TEXT UNIQUE NOT NULL,
-            password_hash TEXT NOT NULL,
-            salt TEXT NOT NULL,
-            role TEXT NOT NULL DEFAULT 'user' 
-                )
-        ''')
-
-        cursor.execute('''
-            CREATE TABLE IF NOT EXISTS images (
-                image_id TEXT PRIMARY KEY,
-                doc_id TEXT NOT NULL,
-                sort_order TEXT,
-                original_path TEXT NOT NULL,
-                processed_path TEXT NOT NULL,
-                changes TEXT,
-                FOREIGN KEY (doc_id) REFERENCES documents (doc_id) ON DELETE CASCADE
-            )
-        ''')
+            self.provider.disconnect()
             
-        # Safely migrate existing databases to include new columns if missing
-        # will need to be expanded to cover other possible changes
-        cursor.execute("PRAGMA table_info(documents)")
-        existing_columns = {col[1] for col in cursor.fetchall()}
-        
-        if 'metadata_json' not in existing_columns:
-            cursor.execute("ALTER TABLE documents ADD COLUMN metadata_json TEXT")
-
-        self.conn.commit()
-    
-    def _create_user(self, username, password,role) -> tuple[bool, str]:
-        """Generates a new user and safely handles duplicate usernames."""
-        cursor = self.conn.cursor()
-        salt = secrets.token_hex(16)
-        hashed_pw = hashlib.pbkdf2_hmac(
-            'sha256', 
-            password.encode('utf-8'), 
-            salt.encode('utf-8'), 
-            100000 
-        ).hex()
-
-        try:
-            cursor.execute('''
-                INSERT INTO users (username, password_hash, salt, role)
-                VALUES (?, ?, ?, ?)
-            ''', (username, hashed_pw, salt, role))
-            
-            self.conn.commit()
-            return True, "User created successfully."
-            
-        except sqlite3.IntegrityError:
-            # The username already exists Roll back the transaction.
-            self.conn.rollback()
-            return False, f"The username '{username}' is already taken."
-            
-        except Exception as e:
-            # Catch any other database errors (like a locked file)
-            self.conn.rollback()
-            return False, f"Database error: {str(e)}"
-
-    def _get_users(self):
-        cursor = self.conn.cursor()
-        cursor.execute("SELECT user_id, username, role FROM users")
-        return [dict(row) for row in cursor.fetchall()]
-
-    def _delete_user(self, user_id):
-        try:
-            cursor = self.conn.cursor()
-            cursor.execute("DELETE FROM users WHERE user_id = ?", (user_id,))
-            if self._has_admin():
-                self.conn.commit()
-                return True, "User deleted."
-            else:
-                raise ValueError("Cannot delete last admin account.")
-        except Exception as e:
-            self.conn.rollback()
-            raise
-
-    def _load_documents(self, filters: dict = None) -> dict[str, Document]:
-        """method for fetching documents. Safely uses parameterized queries with a column whitelist."""
-        docs = {}
-        cursor = self.conn.cursor()
-        
-        sql_query = "SELECT doc_id, metadata, deskewed, needs_approval, approved, rejected, uploaded, path, metadata_file, metadata_file_type, metadata_json, last_modified, error_msg FROM documents"
-        params = []
-        
-        if filters:
-            conditions = []
-        
-            for key, value in filters.items():
-                if key != 'or_filters':
-                    if key not in VALID_FILTERS:
-                        raise ValueError(f"Invalid filter column: {key}")
-                    db_value = 1 if value is True else (0 if value is False else value)
-                    conditions.append(f"{key} = ?")
-                    params.append(db_value)
-            
-            if 'or_filters' in filters:
-                or_conditions = []
-                for or_key, or_value in filters['or_filters'].items():
-                    if or_key not in VALID_FILTERS:
-                        raise ValueError(f"Invalid filter column: {or_key}")
-                    db_value = 1 if or_value is True else (0 if or_value is False else or_value)
-                    or_conditions.append(f"{or_key} = ?")
-                    params.append(db_value)
-                
-                if or_conditions:
-                    or_string = " OR ".join(or_conditions)
-                    conditions.append(f"({or_string})")
-            
-            if conditions:
-                sql_query += " WHERE " + " AND ".join(conditions)
-
-        cursor.execute(sql_query, tuple(params))
-        for row in cursor.fetchall():
-            doc = Document.from_db_row(row)
-            docs[doc.doc_id] = doc
-            
-        if docs:
-            doc_ids = tuple(docs.keys())
-            placeholders = ','.join('?' for _ in doc_ids)
-            
-            image_query = f"SELECT image_id, doc_id, sort_order, original_path, processed_path, changes FROM images WHERE doc_id IN ({placeholders})"
-            cursor.execute(image_query, doc_ids)
-            
-            for row in cursor.fetchall():
-                image_id = row['image_id']
-                doc_id = row['doc_id']
-                order = row['sort_order']
-                orig_path = row['original_path']
-                proc_path = row['processed_path']
-                changes = row['changes']
-                
-                if changes:
-                    changes = json.loads(changes)
-                if doc_id in docs:
-                    docs[doc_id].add_image(image_id, order, orig_path, proc_path,changes)
-        
-        return docs
-
-    def _delete_document(self, doc_id: str):
-        """Internal method. Removes a document and cascades the deletion to its images."""
-        cursor = self.conn.cursor()
-        
-        cursor.execute("DELETE FROM documents WHERE doc_id = ?", (doc_id,))
-        self.conn.commit()
-
-    def _save_document(self, doc: Document):
-        """Internal method. Saves a single Document object."""
-        update_metadata_file(doc)
-        cursor = self.conn.cursor()
-        status = doc.status
-        params_tuple = (
-                doc.doc_id,  
-                int(status.get('metadata', 0)),
-                int(status.get('deskewed', 0)), 
-                int(status.get('needs_approval', 0)),
-                int(status.get('approved', 0)), 
-                int(status.get('rejected', 0)),
-                int(status.get('uploaded', 0)), 
-                str(doc.path), 
-                str(doc.metadata_file) if doc.metadata_file else None, 
-                str(doc.metadata_file_type) if doc.metadata_file_type else None,
-                doc.metadata.to_json() if doc.metadata else None,
-                doc.last_modified, doc.error_msg if doc.error_msg else None
-                )
-
-        cursor.execute('''
-            INSERT OR REPLACE INTO documents (doc_id, metadata, deskewed, needs_approval,
-            approved, rejected, uploaded, path, metadata_file, metadata_file_type, metadata_json, last_modified, error_msg)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        ''',params_tuple )
-        
-        for image_id, image_data in doc.images.items():
-            cursor.execute('''
-                INSERT OR REPLACE INTO images (image_id, doc_id, sort_order, original_path, processed_path, changes)
-                VALUES (?, ?, ?, ?, ?, ?)
-            ''', (image_id, doc.doc_id, image_data['order'], image_data['original'], image_data['processed'], json.dumps(image_data['changes'])))
-        self.conn.commit()
-
-    def _verify_login(self, username, plaintext_password) -> tuple[bool, str | None]:
-        """
-        Hashes the incoming password and compares it to the database.
-        Returns (True, role) if successful, (False, None) if failed.
-        """
-        cursor = self.conn.cursor()
-        cursor.execute("SELECT password_hash, salt, role FROM users WHERE username = ?", (username,))
-        row = cursor.fetchone()
-        
-        # Username doesn't exist
-        if not row:
-            return False, None
-            
-        stored_hash = row['password_hash']
-        salt = row['salt']
-        role = row['role']
-        
-        # Hash the attempted password using the exact same algorithm and salt
-        test_hash = hashlib.pbkdf2_hmac(
-            'sha256', 
-            plaintext_password.encode('utf-8'), 
-            salt.encode('utf-8'), 
-            100000
-        ).hex()
-        
-        # Compare the hashes
-        if test_hash == stored_hash:
-            return True, role
-            
-        return False, None
-
-    def _has_admin(self) -> bool:
-        """Checks if there is at least one admin account in the database."""
-        cursor = self.conn.cursor()
-        cursor.execute("SELECT 1 FROM users WHERE role = 'admin' LIMIT 1")
-        return cursor.fetchone() is not None
-
     def has_admin_sync(self) -> bool:
-        """
-        Synchronously checks if an admin account exists. 
-        Useful for initial startup checks before the worker loop is running.
-        """
-        if not self.db_path.exists():
-            return False
-        try:
-            with sqlite3.connect(self.db_path) as conn:
-                cursor = conn.cursor()
-                cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='users'")
-                if not cursor.fetchone():
-                    return False
-                cursor.execute("SELECT 1 FROM users WHERE role = 'admin' LIMIT 1")
-                return cursor.fetchone() is not None
-        except sqlite3.Error:
-            return False
-
+        return self.provider.has_admin_sync()
+        
     def cancel_current_query(self):
-        """Called by the main GUI thread to instantly abort the active SQLite query."""
-        if self.conn:
-            self.conn.interrupt()
+        self.provider.cancel_current_query()
