+++ source/model/service/__init__.py
@@ -1,5 +1,5 @@
 from .Signals import JobTicket, DatabaseTicket
-from .DatabaseManager import DatabaseManager
+from .DatabaseManager import DatabaseWorker
 from .DocumentPipeline import DocumentPipelineWorker
 #from .ImageLoader import ImageLoader
 from .UploadManager import UploadManager
